+++ src/GUI/GUI.java
@@ -8,7 +8,7 @@ import java.util.Map;
 import com.formdev.flatlaf.FlatDarkLaf;
 import com.formdev.flatlaf.FlatLightLaf;
 
-import GUI.Errors.ErrorDisplay;
+import GUI.Popups.PopupDisplay;
 import GUI.Settings.SettingsGUI;
 import Utils.Utils;
 
@@ -388,7 +388,7 @@ public class GUI {
             isDarkMode = darkMode;
 
         } catch (Exception e) {
-            ErrorDisplay.throwErrorPopup("Es ist ein Fehler beim setzen des Themes aufgetreten.\n" +
+            PopupDisplay.throwErrorPopup("Es ist ein Fehler beim setzen des Themes aufgetreten.\n" +
                     "Das Programm wird möglicherweise etwas anders aussehen als sonst!");
             e.printStackTrace();
         }
