+++ src/GUI/Popups/PopupDisplay.java
@@ -1,11 +1,11 @@
-package GUI.Errors;
+package GUI.Popups;
 
 import javax.swing.JOptionPane;
 
 /*
  * Diese Klasse erstellt eine Fehlermeldung und gibt dies dem Enduser zurück
  */
-public class ErrorDisplay { //TODO: @Leon refactor class name to something more fitting
+public class PopupDisplay { //TODO: @Leon refactor class name to something more fitting
     public static void throwErrorPopup(String errorMsg) {
         JOptionPane.showMessageDialog(null, errorMsg, "Ein Fehler ist aufgetreten", JOptionPane.ERROR_MESSAGE);
     }
