+++ src/Utils/Data/Config/Config.java
@@ -6,7 +6,7 @@ import java.nio.file.FileSystems;
 
 import com.google.gson.Gson;
 
-import GUI.Errors.ErrorDisplay;
+import GUI.Popups.PopupDisplay;
 import Utils.Data.Config.Settings.AppTheme;
 
 public class Config {
@@ -44,7 +44,7 @@ public class Config {
 
             } catch (Exception e) {
                 e.printStackTrace();
-                ErrorDisplay.throwErrorPopup(e.getMessage());
+                PopupDisplay.throwErrorPopup(e.getMessage());
             }
 
         } else {
